Add partitions
Unlike like boost::partition, which partitions a range according to some
existing key, these partitions are composed by partition classes which
are actual C++ classes, which either own or just point to their members.

The members in turn point to their classes. This pointer is automatically
updated when classes are merged.

An implementation on the basis of boost::partition might however be
a good idea, if merges are frequent and resorting is faster then
inserting multiple elements.