Incorporate TreeModel Representation in workflow
It is now implicitly generated and registered at its Entity during a
transformation workflow and can be accessed with getRepresentation().

The BGTree is now a TreeModel -> Blocks Transformation, rather than
Netlist -> Blocks.

Link fixing/repeater placement is now done completely in TreeModel,
without the need to create Blocks. It however still has the methods to
test the links in Blocks via a BlockSim, which is still used by BGTree to
verify that all links are working correctly.