Remove "open" property from interfaces and structure
In e85b86 (Monster refactor mainly structure construction) we already
relieved users/hardware designers from having to avoid open/open
connections, as that was taken care of by the structure2netlist
transformation. But openness was still a property of the Entity's
interface.

An Entity can have multiple representations, even of the same kind, where
the same (ISingle) interface may be actually open in one and closed in the
other. However, we cannot take advantage of that if openness is regarded
a property of the interface. When a representation is added that is open
at an Element that was declared non-open in the Entity's interface, we
would have to either change the interface, or modify that representation
to adhere to it.

Therefore, we finally removed the openness property from the Interface.

It is also not a part of structural representations, since structures are
composed of entities, not concrete representations.