DefaultTransformation: Use factory method for representation creation
TreeModel possibly needs to create itself multiple times due to
unsynthesizable and thus split vertices. To do that in the constructor,
we would have to implement the move constructor, which would be quite
complex due to all the back references. It is easier, less error-prone
and more maintainable to just use a factory method.

The default implementation of the factory method in RepresentationBase
just calls the existing constructor.