Remove awkward C handles for structure::builder::Port
Instead, provide the structure::builder::Ports with a C struct (via
Wrapper<Port>), so they can be directly used from C, without the
unneccessary extra handle layer.

PortHandle, however, which is used for the C++ API, is still required,
since it transparently manages the replacement of a BuilderPort to
an ExistingPort.

Likewise, in order to keep the C structure of a realized BuilderPort at
the same address, Port needs a pointer to it that can be moved to the
ExistingPort. Therefore, Port is not a TypedCObject but a plain CObject
and implements the necessary methods to be used like a TypedCObject.