Make object resolution in the ls command similar to other commands
I.e., don't provide direct access to rhdl_entity_ts, directly displaying
their interface (instead of showing namespaces "interface" and
"representations").

Ls is still different from other commands as it takes a QName and does
the resolution itself, required by the fact that the identifiers of
a currently-defined structure are not available via resolve_*

We might change this by introducing a context namespace that may or may
not be below the root namespace and that can be searched additionally via
resolve_with_bases_* and maybe making that namespace accessible to
Parameter objects, so ls doesn't have to do qn resolution by itself.

We might also loose "complete_object_contextually" if we manage to supply
the context namespace also to ObjectCompleters.

Generalizing the idea of contextually using/importing additional
namespaces, we could also add a use/with/import command, that moves a
namespace in scope, removing the need for full qualification.