Add "help" command
which is a bit more complicated to implement, since it requires a
reference to the other commands (and even those of a fallback).

Therefore, its CmdFn1 help_dummy() does nothing (should maybe panic)
and, together with its exec() function, is never called, since the
control flow is diverted in Processor::exec().

This is not really nice. We should maybe at least use the argument
parsing facility of Cmd1. Maybe we could let help_dummy (after
renaming it) set a flag that indicates that help was requested
(and for what) and evaluate that afterwards. However, most of that
would have to be done in the trait implementation.